requirements.txt
torch
numpy
opencv-python
tqdm
matplotlib
scikit-learn
tensorflow
pillow
scikit-learn
torchvision
ipykernel
seaborn
